refactor(virtqueue): Don't swallow errors on incorrect pop order
Currently, callers have no way to know whether pop_front_* methods
returning None means that no data was received yet or whether they were
trying to pop in the wrong order, e.g. trying to read a packet payload
before the header. This would keep repeating infinitely if a driver was
reading in the wrong order.

This is an irrecoverable programming mistake that we should not swallow,
we should panic instead.

Signed-off-by: Jens Reidel <adrian@travitia.xyz>