refactor(arch/x86_64): Dynamically allocate XSAVE area
The real size of the XSAVE area depends on the enabled features in XCR0.
To avoid having to hardcode the order of the save regions, we can
instead rely on the raw-cpuid crate's APIs that expose the required size
of the area.

However, we cannot *fully* rely on this and just zero out the memory,
since some fields are required to be initialized with non-zero values.
Therefore we need to keep the legacy header definition and manually set
those.

Additionally, this commit introduces support for XSAVEC and XSAVEOPT,
which are both faster and more efficient than XSAVE.

Signed-off-by: Jens Reidel <adrian@travitia.xyz>