fix(arch/aarch64): Restore and save FPU state
Currently, we don't save or restore any FPU state on aarch64, while we
do so on x86_64 and RISC-V.

ARMv8-A mandates NEON support, which adds 32 128-bit vector registers
v0-v32. Of these, only the lower 64 bits of v8-v15 (=d8-d15) are
callee-saved and therefore need to be taken care of by the kernel.

For implementing this, I chose a similar approach to our x86_64 code: We
disable floating point when switching to a new task by setting FPEN in
CPACR_EL1 to 00 so that any use of v0-v32 triggers a trap. We then
handle the trap to enable floating point again and switch the FPU owner
lazily.

There are multiple extensions to ARMv8-A that add more things we need to
save, but for now let's at least implement the baseline properly.

Signed-off-by: Jens Reidel <adrian@travitia.xyz>